refactor generated smoke harness boundaries
