+++ tests/unit/generated-project-smoke-reference-lane.test.ts
@@ -9,23 +9,39 @@ test("generated project smoke script supports a reference example lane", () => {
 		join(repoRoot, "scripts", "run-generated-project-smoke.mjs"),
 		"utf8",
 	);
+	const exampleHelper = readFileSync(
+		join(repoRoot, "scripts", "lib", "generated-project-smoke-example.mjs"),
+		"utf8",
+	);
+	const assertionHelper = readFileSync(
+		join(repoRoot, "scripts", "lib", "generated-project-smoke-assertions.mjs"),
+		"utf8",
+	);
+	const coreHelper = readFileSync(
+		join(repoRoot, "scripts", "lib", "generated-project-smoke-core.mjs"),
+		"utf8",
+	);
 
 	expect(smokeScript).toContain("exampleProject");
 	expect(smokeScript).toContain("--example-project");
+	expect(smokeScript).toContain('./lib/generated-project-smoke-example.mjs');
+	expect(smokeScript).toContain('./lib/generated-project-smoke-assertions.mjs');
 	expect(smokeScript).toContain("runExampleProjectSmoke");
-	expect(smokeScript).toContain("assertExampleProjectScaffold");
-	expect(smokeScript).toContain("ensureCopiedExampleSupportDependencies");
-	expect(smokeScript).toContain("collectProjectFilePaths");
-	expect(smokeScript).toContain("shouldRunMigrationSmoke");
-	expect(smokeScript).toContain('exampleProject === "my-typia-block"');
-	expect(smokeScript).toContain('path.join(projectDir, "build", "blocks", blockSlug)');
-	expect(smokeScript).toContain('devDependencies["bun-types"]');
-	expect(smokeScript).toContain('devDependencies["@types/node"]');
-	expect(smokeScript).toContain(
+	expect(smokeScript).toContain("assertGeneratedProjectScaffold");
+	expect(exampleHelper).toContain("ensureCopiedExampleSupportDependencies");
+	expect(exampleHelper).toContain("runExampleProjectSmoke");
+	expect(exampleHelper).toContain("shouldRunMigrationSmoke");
+	expect(exampleHelper).toContain('devDependencies["bun-types"]');
+	expect(exampleHelper).toContain('devDependencies["@types/node"]');
+	expect(assertionHelper).toContain("assertExampleProjectScaffold");
+	expect(assertionHelper).toContain("collectProjectFilePaths");
+	expect(assertionHelper).toContain('exampleProject === "my-typia-block"');
+	expect(assertionHelper).toContain('path.join(projectDir, "build", "blocks", blockSlug)');
+	expect(coreHelper).toContain(
 		'Expected ${configPath} to declare currentMigrationVersion in a supported format'
 	);
-	expect(smokeScript).toContain('Missing "typecheck" script in');
-	expect(smokeScript).toContain('path.resolve(__dirname, "..", "examples", exampleProject)');
+	expect(exampleHelper).toContain('Missing "typecheck" script in');
+	expect(exampleHelper).toContain('path.resolve(repoRoot, "examples", exampleProject)');
 });
 
 test("CI generated smoke matrix includes the checked-in example lanes", () => {
